requirements/development.txt
# Specifies requirements used in development
-r production.txt
-r test.txt
-r documentation.txt

tox
commitizen
fake-bpy-module-3.0